Reestruturação do projeto, início do relatório e do README
