+++ requirements.txt
@@ -1,0 +1,2 @@
+opencv-python==4.8.1
+ultralytics==8.0.106
