+++ src/point.py
@@ -1,0 +1,11 @@
+class Point:
+    '''
+    Representa um ponto.
+
+    Args:
+        - x (int): Coordenada x do ponto.
+        - y (int): Coordenada y do ponto.
+    '''
+    def __init__(self, x: int, y: int):
+        self.x = x
+        self.y = y
